moves config types
